+++ src/Utils/templates/cloudbuild.yaml.tmpl
